pyproject.toml
[project]
name = "reports"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.64.0",
    "plotly>=6.5.2",
    "python-dotenv>=1.2.1",
    "scipy>=1.17.1",
    "streamlit>=1.54.0",
]
